+++ drift/test/database/types/type_converter_test.dart
@@ -1,20 +1,24 @@
 import 'dart:convert';
 
 import 'package:drift/drift.dart';
+import 'package:drift/extensions/json1.dart';
 import 'package:test/test.dart';
 
 import '../../generated/converter.dart';
 import '../../generated/todos.dart';
+import '../../test_utils/test_utils.dart';
 
 enum _MyEnum { one, two, three }
 
 void main() {
   test('TypeConverter.json', () {
+    // ignore: deprecated_member_use_from_same_package
     final converter = TypeConverter.json(
       fromJson: (json) => _MyEnum.values.byName(json as String),
       toJson: (member) => member.name,
     );
 
+    // ignore: deprecated_member_use_from_same_package
     final customCodec = TypeConverter.json(
       fromJson: (json) => _MyEnum.values.byName(json as String),
       json: JsonCodec(toEncodable: (object) => 'custom'),
@@ -29,10 +33,59 @@ void main() {
     values.forEach((key, value) {
       expect(converter.toSql(key), value);
       expect(converter.fromSql(value), key);
+      expect(converter.toJson(key), value);
+      expect(converter.fromJson(value), key);
 
       expect(customCodec.toSql(key), '"custom"');
       expect(customCodec.fromSql(value), key);
     });
+  });
+
+  test('TypeConverter.json2', () {
+    final converter = TypeConverter.json2(
+      fromJson: (json) => _MyEnum.values.byName(json as String),
+      toJson: (member) => member.name,
+    );
+
+    final customCodec = TypeConverter.json2(
+      fromJson: (json) => _MyEnum.values.byName(json as String),
+      json: JsonCodec(toEncodable: (object) => 'custom'),
+    );
+
+    const values = {
+      _MyEnum.one: ('"one"', 'one'),
+      _MyEnum.two: ('"two"', 'two'),
+      _MyEnum.three: ('"three"', 'three'),
+    };
+
+    values.forEach((key, v) {
+      final (value, jsonValue) = v;
+
+      expect(converter.toSql(key), value);
+      expect(converter.fromSql(value), key);
+      expect(converter.toJson(key), jsonValue);
+      expect(converter.fromJson(jsonValue), key);
+
+      expect(customCodec.toSql(key), '"custom"');
+      expect(customCodec.fromSql(value), key);
+      expect(customCodec.toJson(key), key);
+      expect(customCodec.fromJson(jsonValue), key);
+    });
+  });
+
+  test('TypeConverter.jsonb', () async {
+    final converter = TypeConverter.jsonb(
+      fromJson: (json) => _MyEnum.values.byName(json as String),
+      toJson: (member) => member.name,
+    );
+    final db = TodoDb(testInMemoryDatabase());
+
+    for (final value in _MyEnum.values) {
+      final converted = Variable.withBlob(converter.toSql(value)).json();
+      final query = await db.selectExpressions([converted]).getSingle();
+
+      expect(query.read(converted), '"${value.name}"');
+    }
   });
 
   test('TypeConverter.extensionType', () {